improve test suite spec naming and unit test locations
